Organized the folder structure
