+++ src/main/java/org/example/synchronization/Banking.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.synchronization;
 
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
